requirements.txt
flask
tensorflow-cpu
keras
numpy
gunicorn
